InversionOfControl (IoC) for MonoGameAndGum sample project  Microsoft.Extensions.DependencyInjection with GumService as a service.
